[device-minion-runtime] Increase mbox_message_id_t to keep subsequent kernel_params_t struct members naturally aligned in initial host to MM mailbox messages Print addresses of kernel_params_ptr and kernel_info_ptr passed to launch kernel Create host_message_t struct to clean up handling of messages from host

+++ device-minion-runtime/src/shared/include/mailbox_id.h
@@ -1,7 +1,7 @@
 #ifndef MAILBOX_ID_H
 #define MAILBOX_ID_H
 
-typedef uint32_t mbox_message_id_t;
+typedef uint64_t mbox_message_id_t;
 
 typedef enum
 {
